unified-docker/requirements-nornir.txt
# Core Nornir and NAPALM
nornir==3.5.0
napalm==4.1.0
napalm-ansible==0.1.0

# NAPALM drivers for different vendors
napalm-ios==1.0.0
napalm-iosxr==1.0.0
napalm-nxos==1.0.0
napalm-eos==1.0.0
napalm-junos==1.0.0
napalm-fortios==1.0.0
napalm-panos==1.0.0
napalm-ros==1.0.0
napalm-asa==1.0.0
napalm-ibm==1.0.0
napalm-sros==1.0.0
napalm-vyos==1.0.0

# Nornir plugins
nornir-netmiko==0.1.1
nornir-napalm==0.1.2
nornir-scrapli==2023.1.30
nornir-utils==0.1.2
nornir-jinja2==0.2.0
nornir-rich==0.1.2

# Web framework and API
flask==3.0.0
flask-cors==4.0.0
flask-restx==1.3.0
gunicorn==21.2.0

# Database and data handling
sqlalchemy==2.0.23
alembic==1.13.1
pandas==2.1.4
pydantic==2.5.2

# Configuration and secrets management
python-dotenv==1.0.0
cryptography==41.0.8
keyring==24.3.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Utilities
click==8.1.7
rich==13.7.0
tabulate==0.9.0
jinja2==3.1.2
pyyaml==6.0.1
requests==2.31.0
netaddr==0.10.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Scheduling
apscheduler==3.10.4
celery==5.3.4
redis==5.0.1

# File handling
openpyxl==3.1.2
xlsxwriter==3.1.9

